Support changing the release audit config
Update the workflow to audit past releases with a step that loads the
audit configuration from the GitHub Actions variables so that it can be
changed if necessary, e.g. in case an audit warning is not relevant to
the release (i.e. a new release is not necessary to resolve the warning)