Set up `npm-package-json-lint` and update package metadata (#1555)
* Set up `npm-package-json-lint`

Add `npm-package-json-lint` as a development dependency and use it to
lint the manifest, in particular enforce style and some preference rules
It adds to 1 other tools used for linting the manifest but does provide
some unique features (e.g. one I like is enforcing a certain range style
for different dependency categories).

* Add `bugs` field to the package manifest

Include the standard bugs field with a link where users can report bugs
in the package manifest.

* Use `author` instead of `contributors`

the docs are a bit unclear on what these fields are for exactly, but
from looking around a bit it seems `contributors` is used for active
maintainers more than anything else. Hence, I think using `author` would
be more sensible