Create the `AUTHORS` file
Add a file named `AUTHORS` to the project that lists contributors of the
project, inspired by [1]. This aims to provide more transparency on and
recognition to those that have contributed and thus have licensing right
to the project. Currently, all previous contributors are included and my
intention, in contrast to [1], is to include all future contributors too
unless contributors themselves do not which to be included.

This file is also supported by npm, using it as a fallback for the
manifest's `contributors` field [2] (which this project does not have).

The list was created based on the output of:

    git log --format='%aN <%aE>' | sort -uf

from which bots were excluded and anonymous GitHub emails removed.

--
1. https://opensource.google/documentation/reference/releasing/authors
2. https://docs.npmjs.com/cli/v11/configuring-npm/package-json#default-values