Allow only negative numbers when `allowDerived` is false (#1568)
Update the `no-top-level-side-effects` rule to only allow a unary
expressions at the top level (without `allowDerived`) if it's for a
negative number.

Only minus should be allowed to be able to have a negative number
at the top level, the other unary  operators are (I believe) not necessary.
If you disagree, open an issue and give a reason why the unary
operator is necessary at the top level.