Refactor `no-top-level-side-effects` for readability (#1566)
General refactoring of `no-top-level-side-effects` that aims to simplify
the implementation by reducing duplication and simplifying conditions as
well as whole functions. This refactoring was partially guided by
Codescene.

In the progress of refactoring, two subtle bugs where uncovered. These
were fixed and the test suite was updated to prevent regression. See the
`CHANGELOG.md` for more info.